Refactor index.ts to only export
